pyproject.toml
[project]
name = "rentware-events-backend"
version = "0.1.0"
description = "FastAPI backend for rental events"
authors = [{name = "Rentware"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.23",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic>=2.5",
    "pydantic-settings>=2.2",
    "python-multipart>=0.0.9",
    "email-validator>=2.1",
    "pyjwt>=2.8",
    "passlib[bcrypt]>=1.7",
    "bcrypt>=4.1",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "httpx>=0.25",
    "ruff>=0.1",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
